Application/requirements.txt
Django==1.11.6
psycopg2==2.7.3.1
django-crispy-forms==1.6.1